added lame_encode_flush_nogap().
This routine will flush all internal mp3 buffers, and pad the
last frame with ancillary data so it is a complete mp3 frame.

It will then set ResvSize to 0, so we can continue encoding.

after a call to lame_encode_flush_nogap(), if we close
the mp3 file and open a new one for the future data,
then the two mp3 files should play back 'gapless'
if they are first concatenated together.